Fixed Create page from not loading by adding @model Product at the top.
Added Navigation on the Home Index to redirect to the Index page on the Product controller.

Added EFCore Tools and Design because Enable-Migration / Add-Migration didn't work properly.

Fixed the Create product method by adding the product and saving changes.

Fixed the Delete product method by adding command to remove product and save changes. Also wrapped it in a if statement to check if the product actually exists.